modularize half duplex stream transformer
